backend/src/invoke_chat/requirements.txt
langchain_community
langchain_core
python-dotenv